# gh-121456: fixed complex as string throws error when string contains whitespace