--- a/Misc/NEWS.d/next/Core and Builtins/2024-07-07-10-44-59.gh-issue-121456.ZfD77s.rst
+++ b/Misc/NEWS.d/next/Core and Builtins/2024-07-07-10-44-59.gh-issue-121456.ZfD77s.rst
@@ -1,0 +1,2 @@
+Fixed complex() throws malformed string when using a string with whitespace,
+added more tests for this