# Fixed complex as string throws error when string contains whitespace and added tests for this

issue: gh-121456

Before:
```python
print(complex("12 - 98j"))
```
```
Traceback (most recent call last):
  File "/home/angrygingy/Desktop/temporal/./to_run.py", line 1, in <module>
    print(complex("12 - 98j"))
          ~~~~~~~^^^^^^^^^^^^
ValueError: complex() arg is a malformed string
```

With this changes
```python
print(complex("12 - 98j"))
```
```
(12-98j)
```

I had to add a extra-step to store the sign and later multiply `y`
```c
char sign = *s;
...
if (sign == '-') {
    y = -y;
}
```

My concern was that this affects performance, but I did tests and it doesn't seem to happen.
```python
import cProfile

def example():
    for _ in range(10000000):
        complex("12-98.123456789j")

cProfile.run('example()')
```

Before
```
         4 function calls in 1.648 seconds

   Ordered by: standard name

   ncalls  tottime  percall  cumtime  percall filename:lineno(function)
        1    0.000    0.000    1.648    1.648 <string>:1(<module>)
        1    1.648    1.648    1.648    1.648 show_timings.py:3(example)
        1    0.000    0.000    1.648    1.648 {built-in method builtins.exec}
        1    0.000    0.000    0.000    0.000 {method 'disable' of '_lsprof.Profiler' objects}
```

With this changes
```
         4 function calls in 1.655 seconds

   Ordered by: standard name

   ncalls  tottime  percall  cumtime  percall filename:lineno(function)
        1    0.000    0.000    1.655    1.655 <string>:1(<module>)
        1    1.655    1.655    1.655    1.655 show_timings.py:3(example)
        1    0.000    0.000    1.655    1.655 {built-in method builtins.exec}
        1    0.000    0.000    0.000    0.000 {method 'disable' of '_lsprof.Profiler' objects}
```