@Eclips4 2024-07-07
> `if` statements should follow the format defined in [PEP 7](https://peps.python.org/pep-0007/#code-lay-out)

As well as `while` statements.