@LimaBD 2024-07-07
I had added a check to ensure that the float is not signed twice

```python
print(complex("12 - -98j"))
```
```
Traceback (most recent call last):
  File "/home/angrygingy/Desktop/temporal/./to_run.py", line 1, in <module>
    print(complex("12 - -98j"))
          ~~~~~~~^^^^^^^^^^^^^
ValueError: complex() arg is a malformed string
```