@LimaBD 2024-07-07
I added more tests for twice signed numbers to ensure that ValueError is raised By the way,

why docs tests is not passing ?

```
Error: new NEWS nits:

/home/runner/work/cpython/cpython/build/NEWS:87: WARNING: py:func reference target not found: warnings.filterswarnings
```