@devdanzin 2024-07-07
> By the way, why docs tests is not passing ?

Seems unrelated to your change, it's something added in https://github.com/python/cpython/commit/1a84bdc2371ada60c01c72493caba62c9860007b#diff-3b2105d4c1e6b38172c53e4abe1bcafdce9bb864981c319f6fe15056f6b69d3d